We have web!
Did some integration with the nutrition API and suggestions API. Still
more work needed but a lot of background already implemented.

Close #12
Close #14
Close #17
Close #19
Close #22
Close #26

+++ app/src/main/java/com/felkertech/n/munch/Objects/HistoryItem.java
@@ -9,15 +9,15 @@ import com.felkertech.n.munch.database.FoodTableEntry;
  * Created by N on 3/7/2015.
  */
 public class HistoryItem extends StreamItem {
-    private int calories;
+    private float calories;
     private FoodTableEntry entry;
-    public HistoryItem(String t, String st, int calories, int ri, FoodTableEntry fte) {
+    public HistoryItem(String t, String st, float calories, int ri, FoodTableEntry fte) {
         super(StreamTypes.TYPE_ITEM, t, st, calories+" calories", ri);
         this.calories = calories;
         this.entry = fte;
     }
 
-    public int getCalories() {
+    public float getCalories() {
         return calories;
     }
